Rename back to csgocrate
